Fixed route link mac address plugin.
