+++ vnet/lib/vnet/core/segments/virtual.rb
@@ -22,7 +22,7 @@ module Vnet::Core::Segments
 
     def install
       flows = []
-      flows << flow_create(table: TABLE_TUNNEL_NETWORK_IDS,
+      flows << flow_create(table: TABLE_TUNNEL_IDS,
                            goto_table: TABLE_SEGMENT_SRC_CLASSIFIER,
                            match: {
                              :tunnel_id => flow_tunnel_id
