+++ vnet/lib/vnet/models/mac_lease.rb
@@ -5,7 +5,7 @@ module Vnet::Models
     taggable 'ml'
 
     plugin :paranoia_is_deleted
-    plugin :mac_address_old
+    plugin :mac_address
 
     many_to_one :interface
     one_to_many :ip_leases
