+++ vnet/spec/vnet/endpoints/1.0/shared_examples/many_to_many_relation.rb
@@ -70,20 +70,9 @@ shared_examples "many_to_many_relation" do |relation_suffix, post_request_params
     before(:each) do
       add_relation = "add_#{relation_suffix.chomp("s")}"
 
-      begin
-        entries.times {
-          base_object.send(add_relation, Fabricate(relation_fabricator))
-        }
-      rescue NameError => e
-        pp "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
-
-        pp "add_relation:#{add_relation} relation_fabricator:#{relation_fabricator}"
-        pp "base_object:#{base_object.inspect}"
-
-        e.backtrace.each { |str| pp str }
-
-        raise e
-      end
+      entries.times {
+        base_object.send(add_relation, Fabricate(relation_fabricator))
+      }
 
       get api_relation_suffix
     end
